requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0

# OCR and Image Processing
pytesseract>=0.3.8
Pillow>=8.3.1
opencv-python>=4.5.3

# Machine Learning
scikit-learn>=0.24.2
tensorflow>=2.6.0
torch>=1.9.0

# Document Processing
pdf2image>=1.16.0
python-docx>=0.8.11
openpyxl>=3.0.7

# API and Web Interface
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Development
jupyter>=1.0.0
black>=21.7b0
flake8>=3.9.2
